requirements.txt
dash
python-dotenv 
psycopg2
supabase